requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
openai>=1.10.0,<2.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
pydantic>=2.0.0,<3.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3
schedule==1.2.0
pytest==7.4.3
pytest-asyncio==0.21.1